ATHO-16 ATHO-FR-2 GET_USERS_OF_ROLE tamamlandı.Fail testler düzeltildi.

+++ src/main/java/com/teammental/authorization/service/RoleServiceImpl.java
@@ -2,11 +2,11 @@ package com.teammental.authorization.service;
 
 import com.teammental.authorization.dto.RoleDto;
 import com.teammental.authorization.entity.Role;
+import com.teammental.authorization.exception.RoleException;
 import com.teammental.authorization.jpa.RoleRepository;
 import com.teammental.memapper.MeMapper;
 import java.util.List;
 import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -15,18 +15,21 @@ import org.springframework.stereotype.Service;
  * Created by okan on 5.08.2017.
  */
 @Service
-public class RoleServiceImpl implements   RoleService {
+public class RoleServiceImpl implements RoleService {
 
   //region Messages
   public static final String NOT_FOUND_MESSAGE = "Role kaydı bulunamadı.";
   public static final String PARAMETERS_MUST_BE_NOT_NULL = "Parametre girilmesi gerekmektedir.";
   public static final String ROLE_NAME_MUST_BE_NOT_NULL = "Role Adı girilmesi gerekmektedir.";
-  public static final String ROLE_SHOULD_NOT_HAVE_USERROLE = "Ülkeye ait şehirler bulunmaktadır.";
+  public static final String ROLE_SHOULD_NOT_HAVE_USERROLE = "Role ait kullanıcılar bulunmaktadır.";
   //endregion
 
   //private static final Logger LOGGER = LoggerFactory.getLogger(RoleServiceImpl.class);
   @Autowired
   RoleRepository roleRepository;
+
+  @Autowired
+  RoleUserService roleUserService;
 
   @Override
   public List<RoleDto> getAll() {
@@ -43,16 +46,19 @@ public class RoleServiceImpl implements   RoleService {
     if (entity == null) {
       throw new Exception(NOT_FOUND_MESSAGE);
     }
-
-    return null;
+    Optional<RoleDto> target = MeMapper.getMapperFrom(entity).mapTo(RoleDto.class);
+    return target.get();
   }
 
   @Override
-  public void deleteById(Integer roleId) throws Exception {
+  public void deleteById(Integer roleId) throws RoleException {
     //userrole tablosunda role varsa kontrol yapılacak.
-    //if (cityservice.existCityByCountry(roleId)) {
-    //      throw new Exception(COUNTRY_SHOULD_NOT_HAVE_CITY);
-    //    }
+    if (roleUserService.existUserByRole(roleId)) {
+      throw new RoleException(0, ROLE_SHOULD_NOT_HAVE_USERROLE);
+    }
+    if (!roleRepository.exists(roleId)) {
+      throw new RoleException(1, NOT_FOUND_MESSAGE);
+    }
     roleRepository.delete(roleId);
   }
 
@@ -68,7 +74,7 @@ public class RoleServiceImpl implements   RoleService {
 
     Optional<Role> target = MeMapper.getMapperFrom(dto).mapTo(Role.class);
 
-    Role entity = new Role();
+    Role entity;
     entity = target.get();
     entity = roleRepository.save(entity);
